+++ Base/USTK/Testing/itkLabelMapToAcousticImpedanceImageFilterTest.cxx
@@ -21,11 +21,11 @@ limitations under the License.
 
 =========================================================================*/
 
-#include <itkLabelMapToAcousticImpedanceImageFilter.h>
+#include "itkLabelMapToAcousticImpedanceImageFilter.h"
 #include <fstream>
 
-#include "itkImageFileReader.h"
-#include "itkImageFileWriter.h"
+#include <itkImageFileReader.h>
+#include <itkImageFileWriter.h>
 
 
 template< class TLookupTable >
@@ -126,7 +126,7 @@ int ReadLookupTableFromCSV( const char * filename, TLookupTable & lookupTable )
     return EXIT_FAILURE;
     }
 
-  size_t label = 0;
+  unsigned int label = 0;
   char tissueType[256];
   float acousticImpedance;
   inputStream >> label;
